+++ Source/Apps/StepUpTest/StepUpTest.h
@@ -27,6 +27,6 @@ void StepUpTest_UI_Init(u16* testTime);
 void RunAStepUpTest(void);
 
 //see .c for details
-u16 GetTestParam(const char askString[], u16 min, u16 max, u16 defaultValue, u16 step, char unitString[], u8 fastSpeed);
+u32 GetTestParam(const char askString[], u32 min, u32 max, u32 defaultValue, u32 step, char unitString[], u8 fastSpeed);
 
 #endif /* __STEPUPTEST_H */
